+++ src/renderer/components/Icon/index.tsx
@@ -22,6 +22,7 @@ export type IconName =
   | 'qr'
   | 'question_mark'
   | 'reaction'
+  | 'settings'
   | 'swap_vert'
 
 type Props = {