+++ src/renderer/system-integration/badge-counter.ts
@@ -3,14 +3,27 @@ import { BackendRemote } from '../backend-com'
 import { runtime } from '../runtime'
 
 async function updateBadgeCounter() {
-  const selectedAccountId = window.__selectedAccountId
-  if (selectedAccountId === undefined) {
-    runtime.setBadgeCounter(0)
-  } else {
-    const count = (await BackendRemote.rpc.getFreshMsgs(selectedAccountId))
-      .length
-    runtime.setBadgeCounter(count)
+  const accountIds = await BackendRemote.rpc.getAllAccountIds()
+
+  let combined_count = (
+    await Promise.all(
+      accountIds.map(
+        async accountId =>
+          // Later TODO when we can mute accounts -> don't include muted and not synced
+          (await BackendRemote.rpc.getFreshMsgs(accountId)).length
+      )
+    )
+  ).reduce((previous, current) => previous + current, 0)
+
+  if (!(await runtime.getDesktopSettings()).syncAllAccounts) {
+    if (window.__selectedAccountId) {
+      combined_count = (
+        await BackendRemote.rpc.getFreshMsgs(window.__selectedAccountId)
+      ).length
+    }
   }
+
+  runtime.setBadgeCounter(combined_count)
 }
 
 export const debouncedUpdateBadgeCounter = debounce(
@@ -20,11 +33,11 @@ export const debouncedUpdateBadgeCounter = debounce(
 )
 
 export function initBadgeCounter() {
-  BackendRemote.on('IncomingMsg', accountId => {
-    if (accountId === window.__selectedAccountId) debouncedUpdateBadgeCounter()
+  BackendRemote.on('IncomingMsg', _ => {
+    debouncedUpdateBadgeCounter()
   })
-  BackendRemote.on('ChatModified', accountId => {
-    if (accountId === window.__selectedAccountId) debouncedUpdateBadgeCounter()
+  BackendRemote.on('ChatModified', _ => {
+    debouncedUpdateBadgeCounter()
   })
   // on app startup:
   debouncedUpdateBadgeCounter()